Add Bulk Label GitHub Action (#8)
This GitHub Action runs when an issues event webhook is fired in your GitHub repo. The action checks if there is an array of bulk labels to add [aaa, bbb, ccc] in markdown in the issue's description. The action works with existing labels defined in your GitHub repo; the labeler expects the first 3 characters of the label name (ex. hel for a "help wanted" label). If there is 1 or > bulk labels, the action labels the issue with the specified label(s).